Reload profile files that have changed when redrawing ROI
To update profile values, users have previously needed to restart the application after editing the `profiles` module. Now that users can load profiles through external files, reload any changes in these files to allow the user to make on-the-fly profile modifications.

During initial profile loading, the modification timestamps of any loaded files are saved. When the user redraws the ROI in the GUI, the current file timestamps are checked against the saved timestamps. If any file has been modified, the profile will be completely reloaded to apply all changes in the original order of sub-profiles.