Fix image export to use ROI settings
The addition of settings for sub-images led to a regression in image export, which expects ROI parameters. Add ROI parameters to the `process_file` function and clarify that the existing parameters are for a sub-image rather than an ROI. By including both sub-image and ROI parameters, image export can load sub-images and crop to an ROI. While changing the signature for `process_file`, shift the series, sub-image, and ROI parameters to be optional since they are not needed for all processing types.