Limit single unique label update workaround
The workaround for updating images first displayed with a single unique label value caused an exception when attempting to update a single unique value image through the main Numpy loading function since it loads as read-only. Since this workaround only appears to be necessary for images displayed with `ListedColormap`, limit this workaround to these images (ie label images shown with `DiscreteColormap`).