Sample command script
The `pipelines.sh` shell script initially started as a way to highlight example commands but evolved into fully automated, generalizable scripts without requiring editing. A script with sample commands could still be useful, however, to demonstrate how to perform individual tasks within MagellanMapper or to serve as a template for one's own script. Provide these commands as a separate Bash script, which can be run with manual modifications to paths and picking commands or serve as a list of example commands to copy into another script.