Allow customizing output path during atlas import
Support the `prefix` command-line argument to set an alternate output path when importing atlases. This prefix can be an existing directory, in which case the output filename will remain the same as before. Otherwise, the prefix directory portion will become the output target directory, and the basename will serve as the template for the output filename. In this way, both directory and filename specification are supported. To specify only a directory when it does not exist yet, create it beforehand.