Support resizing during nD rotation
Rotation has the potential to crop images using the default `resize` option, which is False. To support resizing, the output image must be rescaled to the largest shape of the rotated planes. Although the output planes will all have the same shape, generalize finding this shape to accommodate future cases where the shapes may differ, such as in progressive rotations.