Generalize label erosion and edge-aware refinement for alternate symmetric axes
The label erosion and edge-aware refinement steps include a shortcut option of mirroring changes for images presumed to be symmetric. This option has been applied to atlases, which have been assumed to be imported with symmetry along the x-axis.

To generalize this mirroring and ensure underlying symmetry, we now first check the image for symmetry along each axis until finding a symmetric axis. If found and set to mirror, the image will be spit along that axis for erosion and edge-aware refinement. If a symmetric axis is not found, fall back to full image operations, even if set to mirror. Note that the `neg_labels` register setting will need to be set to detect this symmetry since labels may have been inverted to negative numbers during mirroring while remaining otherwise symmetric.