Sample script paths fallback for Bash < 4.3
The sample script uses a library function that makes use of `nameref` available in Bash >= 4.3 for passing arrays by name. Mac does not provide a compatible Bash by default, and other platforms may also use older Bash versions. As a workaround, use indirect expansion instead of nameref to access the variable. Although this approach does not have the advantages of nameref such as writing directly to the variable, only the array contents need to be accessed in this function.