Sample command to view registered atlas on downsampled image
The standard command to view a registered atlas shows the experimental image used for registration, which may be a single channel extracted from a multichannel image such as the autofluorescent channel. To view the downsampled image including all channels, provide another sample command that sets this channel as the main image, sets a prefix to use for the registered files, and removes the `exp.mhd` registered suffix to avoid attempting to load it.