Showing missing keys if key list given when checking S3 bucket size
Generate a data frame of keys that are given in a list when calculating the bucket size but were not found. This function allows one to identify expected by absent keys. Also fix orders of magnitude adjustment for the total converted size. Add the bucket name to the list of found sizes for completeness.