Command-line task to pre-process whole images and save to separate file
Image have typically been pre-procesed as part of a larger process. Occasionally one may wish to pre-process an image to use repeatedly and thus may want to save the pre-processed file. Provide a command-line `proc` task for performing global pre-processing tasks specified in a new ROI profile parameter. Tasks are performed in the specified sequence to the given channels and saved to a separate file for further processing.