Fix out of bounds error when selecting maximum z offset
The 3D visualization truncates the ROI when it is out of the maximum image range, and the 2D viewer ignores x- or y-offsets outside of the image. Viewing a 2D image where the z-offset is set to the maximum slider value leads to an out of bounds error, however. The max slider value has been set to the image size despite 0-based indexing for easy visualization of the image dimensions, relying on an ROI position check before 3D visualization but missing for 2D visualization. The z-offset in a particular triggers this error because the plane is extracted based on the z-value. Refactor this check to adjust ROI position parameters before both 3D and 2D visualization. Alternatively, the `RangeEditor` provides a `high_label` parameter, though this setting appears to be ignored.