Wrapper script to launch the CLI
The run script currently initializes a GUI because of the various VTK/Mayavi imports, which precludes the script from running in a non-graphical environment and adds unnecessary startup cost if the desired task is a command-line function. While the `cli` module can be called directly, it does not take advantage of the environment management in the run script.

Provide a wrapper script that calls the main run script with the GUI turned off. Refactor the run script start either the CLI or GUI in its main function.