Limit image downsampling in Atlas Editor to images loaded through Numpy
Image downsampling is most critical for performance in the Atlas Editor for images loaded through Numpy, where we use memory mapping to reduce memory utilization but at the expense of longer load times especially for images across the z- or y-axis (ie especially loading x-planes). Images loaded through SimpleITK are already loaded into memory, so downsampling should not have as much of a performance improvement while limiting resolution. To reduce the impact of downsampling, limit it to images loaded through Numpy, which should still prevent hanging that occurred when attempting to open the Atlas Editor for high resolution images before downsampling.

Also clarify that the profile parameter for max downsampling sizes refers to edge lengths rather than the shape or number of planes along a given axis. Change the parameter name to `editor_max_sizes` to reflect this distinction from the image shape.