Launch register tasks from CLI
Register tasks specified by the command-line `--register` argument have been accessed by directly calling the `register` module, which in turn calls the CLI to process arguments before performing the register-related task. To simplify this pathway, allow the user to call the CLI directly, which will hand control over to `register.main` to complete the register-related task. Change the sample commands to use `run_cli.py` for register-related tasks to take advantage of the environment management in the run scripts.