Option to customize separators for arguments parsed to dictionaries
The CLI further parses some arguments to dictionaries, using "=" to delimit sub-arguments from their values and "," to delimit values given as a list. Allow these separators to be customized in case arguments or values need to include these characters. Also avoid setting a mutable dictionary as the default parameter value.