Access `plot_2d` tasks directly from CLI
Allow `plot_2d` tasks to be run directly from the CLI when the `--plot_2d` command-line argument is set. Exit the CLI afterward to avoid further processing. Also move the `plot_2d` import statement to the `cli` header for consistency now that the GUI is not opened directly upon this import.