Fallback to matching world info during registration
Occasionally registration fails because of the dreaded `Too many samples map outside moving image buffer` error. Matching world information (spacing, origin, and direction) may overcome this error, presumably because mismatches in the images' coordinate spaces caused them to be too far apart to initialize the registration. Since registered images are assumed to occupy the same coordinate space, provide a fallback where the moving image's world info is overwritten to the same space as that of the fixed object. This approach may also be helpful to avoid mismatches for unclear reasons that may result from rounding, such as appears to occur occasionally in the label overlap measurement filter.