Option to rescale a moving image prior to registration
Although the affine transformation in Elastix will rescale the image, its ability to find the correct scale appears to depend on the initial image initialization. As a way to manually tweak this scaling ahead of time, provide a `rescale` atlas profile option to specify a factor by which to multiply the moving image's spacing.