Use prefix for registered experiment image loading
The command-line set `prefix` flag has been used for loading registered labels and borders images but not yet experiment images. For consistency, use the setting for experiment images. Also wrap loading in a try-except block since `image5d` may have been loaded previously and not be necessary to load here, so failure to load does not need to halt execution.