Change region highlighter color from white to invisible
The ROI Editor highlights regions by overlaying them with a plane that opacifies outside regions with black while leaving the selected region bright. This brightness also opacifies the selected region, however, by overlaying it with a white color. Use NaN in place of this white color so that the overlay for this region is simply invisible, which appears bright in contrast to the now darkened surrounding region.