Fix hang when saving to file that has been opened as a memmap file
Saving to a sub-image may lead to a hang and improperly saved file if the same file has already been opened as a memmap file. This situation may occur if the sub-image including the same offset and size parameters was saved previously. The sub-image parameters cause the corresponding file to be opened, while the stack detector attempts to save a sub-image to this same path. The save operation ends up hanging, and the Python process cannot close.

As a fix, check whether the main image is a memmap file object pointing to the same path to which the sub-image will be saved, in which case the file will not be saved. Alternatively, the memmap file object could be closed, but it is possible that another reference to the memmap exists, which would prevent closing (https://github.com/numpy/numpy/issues/13510). Until a close method is implemented, simply avoid saving and warn the user.

Also use a `with` context manager when saving the file for simplicity.