Integrate `--flip` command-line argument into transformation settings
Replace the `--flip` command-line argument with `--transform rotate=n`, where `n` specifies the number of 90 degree rotations rather than being limited to 0 or 180 degrees. This change also clarifies that the "flip" functionality refers to rotation rather than flipping across a plane. Remove flip/rotation parameters in a number of `register` functions for simplification, defaulting to use `config.transform`. Also, clarify the usage of this parameter in the sample registration command.