Integrate custom pixel display in the ROI Editor
Use the `PixelDisplay` in the ROI Editor to translate values based on downsampling and overlaid labels in the figure footer for overview images. Add an offset parameter to account for the progressive offset in each overview image. This offset can be either a single coordinate or a set of coordinates corresponding to each image in case offsets differ for each overlaid images, such as for overlaid images of different scales. Change the resized labels image back to int to display the more compact and original value since the resizing outputs a float.

Since the x,y values may be different for each overlaid images, display a set of x,y,z coordinates for each image in the `PixelDisplay`. Also, simplify checking whether the coordinates are within bounds of each image.