Fix intensity wraparound for Atlas Editor text boxes
Color/intensity values above 1 in Matplotlib `TextBox` widgets wraparound the value so that the boxes suddenly become dark. Provide a parameter to set an upper limit on hover color/intensity values to clip the intensity at this value. Default to a value just under 1 so that at least some contrast remains for figures with a white background.