Integrate ROI Editor, Atlas Editor, and Mayavi scene in TraitsUI tabbed panel
The ROI Editor and Atlas Editor are Matplotlib figures spawned from the TraitsUI-based ROI Selector and opening in separate windows. Integrating these separate GUIs into a unified interface provides for a cleaner design. It also facilitates using the TraitsUI controls in the ROI Selector to adjust settings for the Matplotlib-based GUIs by allowing the user to see both parts of the graphical interface side-by-side. The ROI Editor already contains some of this integration, such as the table of blobs in TraitsUI reflected the status of flagged blobs in the interface, which would benefit from this unimodal interface.

To integrate the interfaces, group them together in a TraitsUI tabbed panel. Add a custom TraitsUI Matplotlib editor based on a TraitsUI tutorial[^1] except using the Qt5 backend instead, which is already installed for Mayavi support, based on this guide[^2] and updating for Qt5. Move the Mayavi scene into this tab as well.

Creating the Matplotlib figure requires using the object-oriented API rather than the functional `pyplot` API for figure creation and adding sub-plots. Provide an option to pass in an existing figure to the ROI and Atlas Editors.

[^1]: https://docs.enthought.com/traitsui/tutorials/traits_ui_scientific_app.html#making-a-traits-editor-from-a-matplotlib-plot
[^2]: https://github.com/bgriffen/Python3GUITemplate/blob/master/Commonqt4.py