Defer viewer tab initialization until first opening the tab
The Atlas Editor and Mayavi 3D scene are inactive at start up and may not need to be rendered. A tab change listener allows rendering deferral until the given tab is first selected. Add a tab change listener by finding the corresponding Qt tab panel widget and binding a function. This function finds the selected viewer based on an Enum for viewers and the tab order.

Also change the name of the shown tab attribute to the active Matplotlib figure since the attribute is only for these figures and does not include the Mayavi tab. This change also avoids confusion with the new attribute tracked the selected viewer tab.