Fix Windows installer response to Miniconda download prompt
The prompt for downloading and installing Miniconda in the installation Batch script did not take the user input value because the variable expanded before the block. Restructure the script into `GOTO` blocks based on this recommendation[^1]. Change `set /p` to `choice` instead as recommended here[^2] for immediate and tighter control over responses.

Also, check `CONDA_EXE` instead of `CONDA_PREFIX` for a Conda installation as the latter appears to indicate an active environment, whereas the former is set so long as Conda has been initialized previously[^3].

[^1] https://stackoverflow.com/a/55775540/1911852
[^2] https://stackoverflow.com/a/51117474/1911852
[^3] https://github.com/conda/conda/issues/7126