Perform a silent Miniconda installation on Windows with initialization
The default Miniconda installation does not initialize the command prompt, meaning that `conda` commands are not activated unless run through the Anaconda Prompt. The path to finding the `conda` command to perform this initialization is not clearly available either. As a workaround, running the Miniconda installer in silent mode allows specifying an installation path, which can be used afterward to initialize Conda. This workaround also has the benefit of simplifying installation by avoiding a separate third-party graphical step. The MagellanMapper Conda environment can also be set up immediately after Miniconda installation, avoiding the need to rerun the setup script.