Workaround for loss of figure focus in far parts of canvas
Matplotlib figures appear to lose focus sporadically in the lower right parts of the canvas in testing on the Mac when embedded in the PyQt interface. This loss of focus prevents detection of axes during keyboard events such as right arrow presses on a zoomed sub-plot to trigger jumping to the corresponding overview plot. Mouse press events do detect the corresponding axes, however. Since this overview jump functionality is dependent on identifying the axes associated with a keypress, add a jump handler triggered by a right-arrow + left-click event. The right-arrow alone handler has been kept for backward compatibility and in case the figure focus loss issue is later fixed.