Simplify atlas import prefix to specify directory with trailing file separator
The atlas import `prefix` parameter treated the argument as a directory if the directory existed beforehand. As a simplification, specify a directory by using a trailing file separator, which causes `os.path.dirname` to return the full path except this separator. This behavior also avoids the need to manually generate a directory beforehand.