Flip atlas label positioning vertically to keep within sub-plot
Atlas labels' horizontal positions have been flipped when crossing the midline to minimize their chance of overflowing the sub-plot. Keeping labels within the sub-plot has become even more important after the application of the constrained layout since this constraint causes the sub-plot to resize when the label exits the sub-plot into a small gap, such as the gap between the label and figure edge or surrounding slider. Apply similar behavior to the vertical positioning to further encourage the label to stay within the sub-plot.