Fix clearing specified color/intensity values in Atlas Editor
The Atlas Editor color/intensity picker updates all associated Plot Editors with the specified intensity. To avoid triggering this update when changing the text box value through a callback, the specified intensity updater ignored updates if this value matched that of the Plot Editor's current intensity, assumed to indicate that this Plot Editor triggered the update in the first place. Only the first Plot Editor was checked, however, meaning that editing another Plot Editor would trigger setting a specified intensity, and the next edit would use this value rather than a newly picked value. Also, if the first Plot Editor was set to this intensity, specifying this intensity in the text box would not trigger updates to the other Plot Editors.

To avoid leaking this intensity value, reset the specified intensity value on mouse button release. The specified intensity value is only used on mouse press events to set the intensity, so the specified intensity can be cleared afterward. This change also avoids the need to short-circuit specified intensity updates if a Plot Editor's current intensity value matches it or to reset the specified intensity during the button press.