Base overview plot zooming on the ROI size
The ROI Editor overview zoom plots progressively zoom into the ROI. For large ROIs relative to the image shape, the zoom may overshoot into the ROI. Rather than zooming by a simple power function, limit the zoom range within this power function by the ratio of the image to a minimum size based on the ROI shape. Initially set this minimum size to 3x the ROI shape so that the final overview image still provides context to the ROI.

For large ROIs relative to the image size, this minimum size may exceed the dimensions of the original image, in which case the ROI shape itself is used as the minimum size to maximize the zoom range across overview images. To avoid zooming out, replace zoom values below 1 with a linear range from 1 to the lowest calculated value above 1. Also, change the 3-zoom layout to automatically calculate these zooms rather than specifying them explicitly.