Default to saving figures as PNG files
The figure save function has simply skipped saving if no extension is given. Change this behavior to default to PNG format to ensure file saving. Provide feedback if the given extension is a 3D format before skipping the save. Also, provide an option to save the given rather than the currently active figure.