Save Atlas Editor figures
Previously the Atlas Editor figures have been saved through the Matplotlib toolbar, but we have not included this toolbar in the new integrated UI. A dedicated save function would also allow specifying the filename based on the image parameters. Add a function to save the figure based on the current offset. Update the offset attribute when updating coordinates. Allow the ROI path to lack a size parameter.