Allow selecting multiple regions
The region selector has permitted a single label or positive and negative versions of the label to account for corresponding labels on opposite hemispheres, for example. Occasionally regions from separate super-structures need to be combined, however. Extend this selection functionality to accept multiple distinct labels as a comma-delimited list. These regions and all sub-regions of each label will be combined for the region be highlighted and region measurements.