Change profiles delimiter to comma
Profile names have been delimited by underscores, but the addition of YAML paths as profile names meant that these paths could not include any underscores. Change this delimiter to a comma as a character less likely to be included in paths but not requiring quoting. This delimiter is also more consistent with the delimiter used for other command-line parameters here.