Fix potential for ROI to exceed image boundaries during 3D rendering at a label
When showing a 3D rendering of the raw intensity image with an ROI centered on a given label, the ROI boundaries may exceed the image, leading to a crash. Check the ROI position to adjust the ROI prior to rendering the 3D view. In the ROI check, also check for negative offsets since they may have been set outside of the GUI sliders, such as during this label centering.