Auto-apply dark theme when in dark mode
Qt automatically applies dark mode when the operating system is in dark mode, at least on testing with the Pip-installed PyQt 5.14 on MacOS (not applied in PyQt 5.12.3 from Conda), but Matplotlib figures do not match this theme. To match themes automatically, infer whether the Qt window is in dark mode based on the application window's maximum RGB value as inspired by here[^1] since a Qt API for this detection does not appear to exist currently[^2]. If in dark mode, apply the theme if no other theme has been added by the user. Move Matplotlib style and theme setup to the `Visualization` initialization to avoid redundant setup.

[^1] https://github.com/kovidgoyal/calibre/commit/9fa545f5d9570604a63a00ae0f439d9e94a93f40
[^2] https://bugreports.qt.io/browse/QTBUG-74210