Add a profiles panel
A profiles panel allows the user to select profiles graphically rather than having to specify them at command-line. This panel would also allow users to adjust the combination of profiles to include and could load profile updates from YAML files.

This update moves the ROI Selector into a tabbed pane of options, with the ROI selection panel still showing first and a new profiles panel. The profile selectors consist of dropdowns for the profile category, name, and channel, which get added to a table when the user presses an add button. A load button resets the loaded profiles based on the given items. Each profile category loads in the list of available profiles in the given category as well as profile YAML files. Currently the channel value is used only for ROI profiles.