Fix profile table row deletion
Row deletion in the TraitsUI `TabularAdapter` requires its Trait to be a mutable object, but this object has been a Numpy array, which does not allow row deletion without creating a new array. Extending the `delete` function and deleting the Numpy array row does remove this row in the graphical table, but the underlying array array still appears to point to the original object. Loading profiles thus continues to load the deleted row, and adding a profile triggers the deleted row to return in the table.

To delete rows fully, change the Trait from `Array` to `List` to use a generic Python list. This list will be temporarily converted to a Numpy array simply for fancy indexing to group profiles by category and channel. New profile entries can be added directly to the initial empty list rather than having to generate a Numpy array with `empty` or to copy the array whenever new entries are added.