Pause the setup Batch script so users can see instructions after install
By default the Command Prompt window closes when the Batch script completes if launched by double-clicking. This closing prevents the user from viewing any final instructions from the setup script such as how to activate the Conda environment. Add a `pause` call at the end of the script to keep the window open under the user presses a key to close it.