Allow image brightness reduction
Shift the brightness slider range so that its default value is 0, centered on a range from the negative to the positive 1/5th of the image's maximum data value. By centering the slider, the user can decrease brightness in addition to increasing it. Reducing the range to a fraction of the data type allows more precise control over slider movements.