Adjust resolution for HiDPI displays on Windows and Linux
Although MacOS implements the HiDPI mode that Qt supports natively[^1], Qt has a fallback to support HiDPI on other platforms such as Windows and Linux. This fallback requires setting a HiDPI attribute, which improves resolution in testing on a HiDPI display in Windows 10 and WSL via MobaXterm with X11 scaling turned to "auto". The setting does not appear to affect display in MacOS in testing on a non-HiDPI and a HiDPI screen.

[^1] https://doc.qt.io/qt-5/qt.html