Generalize resizing overlaid images to shape of main image
The ROI Editor has resized overlaid images to the shape of the main image to allow overlaying labels registered to a downsampled image on top of the full resolution image. Generalize this method by moving this resizing functionality to the `plot_support.overlay_images` function so that the Atlas Editor can incorporate this functionality. The Atlas Editor `PlotEditor`s have downsampled each overlaid image by the the same amount, but smaller images do not need as much downsampling. Calculate downsampling on a per-image basis to avoid over-downsampling smaller images. Also calculate scaling to extract the z-plane in each over laid image corresponding to that of the main image.

Now that the non-main images passed to the pixel display have not yet been resized, scale coordinates to these non-main images to account for any resizing before finding corresponding pixel values to the main image's coordinates. Also, rename `z` to `px` to clarify that the value is not the z-index, but the value of the given pixel.