Option to auto-zoom into a region in a `PlotEditor`
Provide a function to zoom into a sub-image at the given offset and size with an editor plot. Since the z-value is not needed, the offset and size consist of only xy values. Add an option in the coordinate transformation function to transform any given range of values rather than always skipping the first value as has been done for 3D coordinates where the z-value was skipped.