Generalize image settings synchronization
Create a mixin class for image synchronization including extracting and setting display values for `PlotEditor`s within a given viewer. Move the synchronization specification from `AtlasEditor` to this class and inherit this class in both `AtlasEditor` and `ROIEditor`. Extend image display adjustment through the `Visualization` GUI to the ROI Editor.