Split out file setup from directory import
The directory image import function has both parsed paths to import and performed the import itself. To allow the user to inspect these paths before import, split out the path parsing into a separate function that outputs a dictionary of paths by channel. Provide an option for all callback function for output display in addition to the terminal during import. Make the `prefix` parameter mandatory to set an output path since the desired name can be ambiguous. No longer default to outputting the file to the image directory, which prevented re-importing the directory.