Image import panel in GUI
Add a panel for interactively importing images through the GUI. This panel splits import into two steps: 1) setting up paths and 2) importing image files at these paths with the given metadata. The user can selected a single image or directory from a file browser, which populates a table of paths and channels related to the chosen path. The table allows modification so that the user can adjust files to import. After selecting files, the user can enter microscopy related image metadata and initiate import. Import runs in a separate thread to allow the GUI to remain interactive including display of the import progress.