Change command-line `channel` parameter to take multiple channels
The channel parameter has only accepted all or a specific channel, but users frequently need to select a subset of channels. To provide this option, change the `channel` parameter to take any number of arguments. For simplicity, no longer accept `-1` as an alias for all channels since all channels are implied by not setting the parameter.

Given the number of functions that use this channel parameter, update those function to accept a sequence of channels. In the ROI verifier, change the ROI profile to be the first listed profile rather than the profile corresponding to the channel index on the assumption that only one profile can be chosen, and the user would only list profiles for selected channels. When exporting ROIS, only use the first listed channel. Setting up channels for an ROI returns only the first given channel if the ROI is single channel.