Accept an arbitrary number of channels when adding blobs
Adding blobs in the ROI Editor has defaulted to channel 0 and used `ctrl+1` to indicate channel 1 but could not go further. To accept arbitrarily high channel numbers, used regex to accept an n-th key (up through 9 as a ctrl key combination). Simply avoid adding the blob if the selected channel is out of range for the given number of channels.