Select channel subsets in the GUI
The GUI has used a slider to select a single channel, with `-1` serving as an alias for all channels. Now that `config.channel` and associated functions have been updated to accept a list of specific channels, incorporate this channel subsetting into the GUI through checkboxes. This change requires updating and synchronizing channel selectors in all panels.

- Change the main channel selector from a slider to check boxes, dynamically updated for the number of channels in the current image
- Similarly replace the profiles channel selector, using the same set of check boxes as for the main channel selector
- Option to the same profile for multiple channels by selected the desired profile channel checkboxes
- Use this same set of channels for the image adjustment channel selector but replace the slider with a dropdown box to select a single channel at a time