Reload the list of profile names when loading profiles
As a way to access newly saved profile files, reload the list of profile names when loading profiles. This behavior requires the user to load profile even before adding the final list, but this loading generally take little time, and "load profiles" appears intuitive with loading a new list of profiles in addition to loading profiles listed in the table.