Ensure that the currently activated Conda environment name starts with "mag"
The run script's check for a currently activated Conda environment has only looked for the presence of any activated environment rather than for the MagellanMapper's default environment. This behavior allows the flexibility of using the run script in a non-default environment, but it also causes the default Conda environment, "base", to appear as an appropriate environment. Since this base environment is activated by default in standard Conda installations, the name of the activated environment should be checked to ensure that it is appropriate for MagellanMapper. To continue to allow at least a little flexibility in the naming, ensure that the environment name begins with "mag" but allow any following characters.

Also, change the environment variable checked to `CONDA_DEFAULT_ENV` since the base environment's `CONDA_PREFIX` value appears to give the Conda install directory rather than the base name.