Import RAW files
RAW files can be loaded directly from Numpy but require knowledge of basic image metadata such as shape and data type. If an image cannot be opened by Bioformats, fall back to opening the image by Numpy if possible, using the `memmap` function to avoid loading the whole file into memory. Add a `shape` parameter to specify the image shape rather than inferring it from the channels dictionary and first image since the RAW file needs to be opened with this shape known a priori. Change the name of these `bioformats` functions to `multipage` since the RAW files are not loaded via Bioformats.