Add an output image shape array field to the import panel
Output images such as RAW files require a manually specified shape parameter. Add a row in the import panel for this array, grouping it with the output image prefix as output parameters. For simplicity, use Numpy types rather than Traits for the data type specification in the `Array` image import variables since they are already Traits and as given in the Traits docs[^1]. Also, change these variables' assignments to assign the full 2D array rather than only the first element so that the value updates on subsequent assignment. Generate via Numpy functions to avoid redundancy and since Traits sequences are Numpy arrays. Clear the resolution and shape values when clearing the import files.

[^1] https://docs.enthought.com/traitsui/traitsui_user_manual/factories_basic.html#arrayeditor