Auto-populate image import metadata fields when available
Metadata extraction has taken place during file import, meaning that the user cannot always verify the metadata beforehand and may not know if the metadata will be automatically extracted or not. To show this metadata, move up extraction from the import step to a separate step after file setup and before import. This extraction requires Bioformats, which causes a delay for Java initialization, but subsequent Java calls no longer need this initialization, so overall performance impact should be minimal. To remain responsive during this initialization, add a thread for collecting this metadata, using a `QThread` to avoid a Java exception when emitting a signal and allowing the signal to be defined within the class.

Additional changes:
- Disable the import button until setup is fully complete to avoid attempting while the setup thread is extracting metadata
- Add a "default" byte order for cases where the byte order is not applicable or default
- Pass metadata as a dictionary to the multipage file import function rather than storing as `config` attributes to avoid overwriting these settings for the current image if the import fails and does not load a new image
- Add shape and data type metadata keys to store this information during import setup