Fix updating imported image shape for user-updated paths
Separating out shape determination from the import step in commit 47c19b94 caused a regression where user updates to the table of paths to import caused the incorrect channels to be read. Determining the output image shape for channels has been complicated by the possibility of either multichannel single files or multiple single-channel files. To distinguish these scenarios, use a special key to designate the former case. This key will be applied only when no filenames with channel designators are found, in which case only the given file will be taken, no longer falling back to any matching file except extension.

This change avoids the ambiguity from relying on the number of channels in the dictionary since the user may reduce the number of channels to a single channel, and it would be ambiguous whether this file has multiple channels or not. To further reduce potential conflicts between these two scenarios, reduce the dictionary to only this item if the key exists. Similarly, prune the dictionary to any channels parameter if the dictionary is assumed to only contain single channel files.